extracted enum DeploymentState
