+++ AMW_business/src/main/java/ch/puzzle/itc/mobiliar/business/deploy/entity/NodeJobEntity.java
@@ -20,26 +20,14 @@
 
 package ch.puzzle.itc.mobiliar.business.deploy.entity;
 
-import java.util.Arrays;
-import java.util.List;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-import javax.persistence.TableGenerator;
-
 import ch.puzzle.itc.mobiliar.business.database.control.Constants;
-import ch.puzzle.itc.mobiliar.business.deploy.entity.DeploymentEntity.DeploymentState;
 import ch.puzzle.itc.mobiliar.common.exception.DeploymentStateException;
 import lombok.Getter;
 import lombok.Setter;
+
+import javax.persistence.*;
+import java.util.Arrays;
+import java.util.List;
 
 @Entity
 @Table(name = "TAMW_NODEJOB")
